requirements.txt
# ═══════════════════════════════════════════════════
# Fast Drop — Requirements (Egyptian Arabic AI System)
# Python 3.13 compatible
# ═══════════════════════════════════════════════════

# Core API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
python-multipart>=0.0.9

# LLM Providers
groq>=0.9.0
google-generativeai>=0.8.0
openai>=1.40.0               # OpenRouter uses OpenAI-compatible SDK
huggingface-hub>=0.25.0      # HuggingFace Inference API (Qwen, Command-R+)

# Arabic NLP
# camel-tools not available for Python 3.13 — regex-based fallback is used
langdetect>=1.0.9

# Embeddings & RAG
sentence-transformers>=3.1.0

# Route Optimization
ortools>=9.10.4067
scikit-learn>=1.5.2
geopy>=2.4.1
numpy>=1.26.4

# Database & Async
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.29.0
alembic>=1.13.1
redis>=5.0.8
celery>=5.4.0

# Data & HTTP
pandas>=2.2.3
httpx>=0.27.2
tabulate>=0.9.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Telegram Bot
python-telegram-bot>=21.6

# Geocoding
googlemaps>=4.10.0

# Arabic Text Rendering
arabic-reshaper>=3.0.0
python-bidi>=0.6.0

# ═══════════════════════════════════════════════
# [7] LlamaIndex & LangSmith Hybrid Stack
# ═══════════════════════════════════════════════
llama-index>=0.11.0
llama-index-retrievers-bm25>=0.2.0
llama-index-embeddings-huggingface>=0.3.0
langsmith>=0.1.100
langchain-huggingface>=0.1.0
llama-index-core>=0.14.0
PyStemmer>=2.2.0
